requirement.txt
pandas
numpy
os
re
uvicron
fastapi
shutil
typing
langchain
langchain_groq
python-dotenv
langchain-core
groq
